+++ lib/generators/v3/index.js
@@ -3,13 +3,13 @@ const Chance = require('chance').Chance();
 const Format = require('./format');
 const Randexp = require('randexp').randexp;
 
-const mock = (schema, useExample)  => {
+const mock = (node, useExample)  => {
     let mock;
 
     // Parameters and responses live inside the schema keyword now
-    if (schema.schema) {
-        schema = schema.schema;
-    }
+    const schema = node.schema;
+
+    console.log(schema)
 
     if (schema) {
         let type = schema.type || findType(schema);
