+++ lib/index.js
@@ -149,7 +149,7 @@ const mockResponse = (response, { version, useExamples }) => {
 
 
 const appropriateGenerators = version => {
-    if (version.match(new RegExp('/^3\./'))) {
+    if (version.match(version.match(/^3\./))) {
         return Generators3;
     } else {
         return Generators2;
@@ -157,7 +157,7 @@ const appropriateGenerators = version => {
 }
 
 const appropriateParamTypes = version => {
-    if (version.match(new RegExp('/^3\./'))) {
+    if (version.match(version.match(/^3\./))) {
         return ParamTypes3;
     } else {
         return ParamTypes2;
