+++ package.json
@@ -8,7 +8,7 @@
   },
   "scripts": {
     "lint": "eslint lib",
-    "test": "mocha tests/*.js",
+    "test": "mocha tests/**/*.js",
     "cover": "istanbul cover _mocha -- tests/*.js"
   },
   "repository": {
